% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/createData.R
\name{createData}
\alias{createData}
\title{Create body-tail data object}
\usage{
createData(y, p, dist = "GH")
}
\arguments{
\item{y}{Transformed data}

\item{p}{Confidence level for risk measures}

\item{dist}{Body distribution for the model}
}
\description{
Create body-tail data object
}

