% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/priors.R
\name{createExpertPrior}
\alias{createExpertPrior}
\title{Creates expert prior from ellicitation}
\usage{
createExpertPrior(m, q, mDiff, qDiff, y)
}
\arguments{
\item{m}{Median for 1\% value at risk}

\item{q}{90\% quantile for median m}

\item{mDiff}{Median for difference between 1\% and 0.1\% value at risk}

\item{qDiff}{90\% quantile for mDiff}

\item{y}{Transformed data}
}
\description{
Creates expert prior from ellicitation
}

