% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/createData.R
\name{createGPData}
\alias{createGPData}
\title{Create fixed-threshold GP data object}
\usage{
createGPData(y, p, gp.p)
}
\arguments{
\item{y}{Transformed data}

\item{p}{Confidence level for risk measures}

\item{gp.p}{Percentile for threshold}
}
\description{
Create fixed-threshold GP data object
}

