% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/createModel.R
\name{createGPModel}
\alias{createGPModel}
\title{Create fixed-threshold GP model}
\usage{
createGPModel(gp.prior)
}
\arguments{
\item{gp.prior}{GP prior}
}
\description{
Create fixed-threshold GP model
}

