% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/createModel.R
\name{createModel}
\alias{createModel}
\title{Create body-tail model}
\usage{
createModel(data, gp.prior)
}
\arguments{
\item{data}{Data object from "createData"}

\item{gp.prior}{Tail-GP prior}
}
\description{
Create body-tail model
}

