% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/filterVolatility.R
\name{filterVolatility}
\alias{filterVolatility}
\title{Volatility filter of transformed data}
\usage{
filterVolatility(transformedData)
}
\arguments{
\item{transformedData}{Transformed data}
}
\description{
Volatility filter of transformed data
}

