% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utilities.R
\name{fitGP_MLE}
\alias{fitGP_MLE}
\title{Fixed-threshold GP fitted using MLE}
\usage{
fitGP_MLE(y, gp.p, gp.u = -1)
}
\arguments{
\item{y}{Transformed data}

\item{gp.p}{Percentile}

\item{gp.u}{Threshold (optional, percentile ignored if defined)}
}
\description{
Fixed-threshold GP fitted using MLE
}

