% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/fitMCMC.R
\name{fitMCMC}
\alias{fitMCMC}
\title{Create fixed-threshold GP data object}
\usage{
fitMCMC(model, d, ivAndCov)
}
\arguments{
\item{model}{Model object}

\item{d}{Data object}

\item{ivAndCov}{Starting values and covariance}
}
\description{
Create fixed-threshold GP data object
}

