% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/generateIV.R
\name{generateIV}
\alias{generateIV}
\title{Generate initial values and covariance matrix}
\usage{
generateIV(model, d, n = 500)
}
\arguments{
\item{model}{Model object}

\item{d}{Data object}

\item{n}{Number of guesses, defaults to 500}
}
\description{
Generate initial values and covariance matrix
}

