% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utilities.R
\name{MRLP}
\alias{MRLP}
\title{Function for mean residual life plot}
\usage{
MRLP(transformedData)
}
\arguments{
\item{transformedData}{Transformed data}
}
\description{
Function for mean residual life plot
}

