% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utilities.R
\name{percentageFromThreshold}
\alias{percentageFromThreshold}
\title{Utility function for calculating percentile from threshold}
\usage{
percentageFromThreshold(u, y)
}
\arguments{
\item{u}{Threshold}

\item{y}{Transformed data}
}
\description{
Utility function for calculating percentile from threshold
}

