% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateGHGP}
\alias{simulateGHGP}
\title{Creates expert prior from ellicitation}
\usage{
simulateGHGP(n = 1000)
}
\arguments{
\item{n}{Sample size, defaults to 1000}
}
\description{
Creates expert prior from ellicitation
}

