% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utilities.R
\name{thresholdFromPercentage}
\alias{thresholdFromPercentage}
\title{Utility function for calculating threshold from percentile}
\usage{
thresholdFromPercentage(p, y)
}
\arguments{
\item{p}{Percentile}

\item{y}{Transformed data}
}
\description{
Utility function for calculating threshold from percentile
}

