% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utilities.R
\name{transformStockData}
\alias{transformStockData}
\title{Utility function for transforming stock data}
\usage{
transformStockData(orgData)
}
\arguments{
\item{orgData}{Non-transformed data}
}
\description{
Utility function for transforming stock data
}

